Testing moving namespaces
