+++ psalm.xml
@@ -56,6 +56,7 @@
                 <file name="src/Bundle/Controller/RequestConfiguration.php" />
                 <file name="src/Bundle/Controller/ResourceController.php" />
                 <file name="src/Bundle/Grid/Parser/OptionsParser.php" />
+                <file name="src/Component/Symfony/Request/ParametersParser.php" />
             </errorLevel>
         </InternalMethod>
 
@@ -170,6 +171,13 @@
             </errorLevel>
         </TooManyArguments>
 
+        <TypeDoesNotContainType>
+            <errorLevel type="suppress">
+                <directory name="src/Bundle/Controller" />
+                <directory name="src/Bundle/Provider" />
+            </errorLevel>
+        </TypeDoesNotContainType>
+
         <UndefinedClass>
             <errorLevel type="suppress">
                 <referencedClass name="Doctrine\Bundle\MongoDBBundle\DependencyInjection\Compiler\DoctrineMongoDBMappingsPass" />
@@ -185,5 +193,11 @@
                 <referencedClass name="UnitEnum"/>
             </errorLevel>
         </UndefinedDocblockClass>
+
+        <UnrecognizedStatement>
+            <errorLevel type="suppress">
+                <file name="src/Bundle/Controller/ParametersParserInterface.php" />
+            </errorLevel>
+        </UnrecognizedStatement>
     </issueHandlers>
 </psalm>
