+++ src/Bundle/Grid/View/ResourceGridViewFactory.php
@@ -13,12 +13,12 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\Grid\View;
 
-use Sylius\Bundle\ResourceBundle\Controller\ParametersParserInterface;
 use Sylius\Bundle\ResourceBundle\Controller\RequestConfiguration;
 use Sylius\Component\Grid\Data\DataProviderInterface;
 use Sylius\Component\Grid\Definition\Grid;
 use Sylius\Component\Grid\Parameters;
 use Sylius\Component\Resource\Metadata\MetadataInterface;
+use Sylius\Component\Resource\Symfony\Request\ParametersParserInterface;
 
 final class ResourceGridViewFactory implements ResourceGridViewFactoryInterface
 {
